Split DiEvent parameters/data blocks into their own files
