+++ Source/SharpNeedle/SonicTeam/DiEvent/Node.cs
@@ -71,7 +71,7 @@ public class Node : IBinarySerializable<GameType>
 
             case NodeType.CharacterMotion:
             case NodeType.ModelMotion:
-                Data = new MotionModelData(reader, game);
+                Data = new ModelMotionData(reader, game);
                 break;
 
             case NodeType.Attachment:
