+++ Source/SharpNeedle/SonicTeam/DiEvent/Parameters/AuraParam.cs
@@ -1,0 +1,92 @@
+namespace SharpNeedle.SonicTeam.DiEvent;
+
+public class AuraParam : BaseParam
+{
+    public int Field00 { get; set; }
+    public int Field04 { get; set; }
+    public int Field08 { get; set; }
+    public int Field0C { get; set; }
+    public float Field10 { get; set; }
+    public float Field14 { get; set; }
+    public float Field18 { get; set; }
+    public float Field1C { get; set; }
+    public float Field20 { get; set; }
+    public int Field24 { get; set; }
+    public int Field28 { get; set; }
+    public int Field2C { get; set; }
+    public int Field30 { get; set; }
+    public float Field34 { get; set; }
+    public float Field38 { get; set; }
+    public float Field3C { get; set; }
+    public float Field40 { get; set; }
+    public float Field44 { get; set; }
+    public int Field48 { get; set; }
+    public float[] ValuesTimeline { get; set; } = new float[32];
+
+    public AuraParam() { }
+    public AuraParam(BinaryObjectReader reader, GameType game)
+        => Read(reader, game);
+
+    public override void Read(BinaryObjectReader reader, GameType game)
+    {
+        Field00 = reader.Read<int>();
+        Field04 = reader.Read<int>();
+        Field08 = reader.Read<int>();
+        Field0C = reader.Read<int>();
+        Field10 = reader.Read<float>();
+        Field14 = reader.Read<float>();
+        Field18 = reader.Read<float>();
+        Field1C = reader.Read<float>();
+        Field20 = reader.Read<float>();
+        Field24 = reader.Read<int>();
+        Field28 = reader.Read<int>();
+        Field2C = reader.Read<int>();
+        Field30 = reader.Read<int>();
+        Field34 = reader.Read<float>();
+        Field38 = reader.Read<float>();
+        Field3C = reader.Read<float>();
+        Field40 = reader.Read<float>();
+        Field44 = reader.Read<float>();
+        Field48 = reader.Read<int>();
+        reader.ReadArray<float>(32, ValuesTimeline);
+    }
+
+    public override void Write(BinaryObjectWriter writer, GameType game)
+    {
+        writer.Write(Field00);
+        writer.Write(Field04);
+        writer.Write(Field08);
+        writer.Write(Field0C);
+        writer.Write(Field10);
+        writer.Write(Field14);
+        writer.Write(Field18);
+        writer.Write(Field1C);
+        writer.Write(Field20);
+        writer.Write(Field24);
+        writer.Write(Field28);
+        writer.Write(Field2C);
+        writer.Write(Field30);
+        writer.Write(Field34);
+        writer.Write(Field38);
+        writer.Write(Field3C);
+        writer.Write(Field40);
+        writer.Write(Field44);
+        writer.Write(Field48);
+        writer.WriteArrayFixedLength(ValuesTimeline, 32);
+    }
+
+    public override int GetTypeID(GameType game)
+    {
+        switch (game)
+        {
+            case GameType.Frontiers:
+                return (int)FrontiersParams.Aura;
+
+            case GameType.ShadowGenerations:
+                return 0;
+
+            default:
+                return 0;
+        }
+    }
+}
