+++ Source/SharpNeedle/SonicTeam/DiEvent/Parameters/CameraShakeLoopParam.cs
@@ -1,0 +1,46 @@
+namespace SharpNeedle.SonicTeam.DiEvent;
+
+public class CameraShakeLoopParam : BaseParam
+{
+    public int Field00 { get; set; }
+    public int Field04 { get; set; }
+    public float Field08 { get; set; }
+    public float Field0C { get; set; }
+    public float Field10 { get; set; }
+    public float Field14 { get; set; }
+    public float Field18 { get; set; }
+    public float Field1C { get; set; }
+    public float[] CurveData { get; set; } = new float[64];
+
+    public CameraShakeLoopParam() { }
+    public CameraShakeLoopParam(BinaryObjectReader reader, GameType game)
+        => Read(reader, game);
+
+    public override void Read(BinaryObjectReader reader, GameType game)
+    {
+        Field00 = reader.Read<int>();
+        Field04 = reader.Read<int>();
+        Field08 = reader.Read<float>();
+        Field0C = reader.Read<float>();
+        Field10 = reader.Read<float>();
+        Field14 = reader.Read<float>();
+        Field18 = reader.Read<float>();
+        Field1C = reader.Read<float>();
+        reader.ReadArray<float>(64, CurveData);
+    }
+
+    public override void Write(BinaryObjectWriter writer, GameType game)
+    {
+        writer.Write(Field00);
+        writer.Write(Field04);
+        writer.Write(Field08);
+        writer.Write(Field0C);
+        writer.Write(Field10);
+        writer.Write(Field14);
+        writer.Write(Field18);
+        writer.Write(Field1C);
+        writer.WriteArrayFixedLength(CurveData, 64);
+    }
+
+    public override int GetTypeID(GameType game) { return (int)ParameterType.CameraShakeLoop; }
+}
