+++ Source/SharpNeedle/SonicTeam/DiEvent/Parameters/CameraShakeParam.cs
@@ -1,0 +1,43 @@
+namespace SharpNeedle.SonicTeam.DiEvent;
+
+public class CameraShakeParam : BaseParam
+{
+    public int Field00 { get; set; }
+    public int Field04 { get; set; }
+    public float Field08 { get; set; }
+    public float Field0C { get; set; }
+    public int Field10 { get; set; }
+    public int Field14 { get; set; }
+    public int Field18 { get; set; }
+    public int Field1C { get; set; }
+
+    public CameraShakeParam() { }
+    public CameraShakeParam(BinaryObjectReader reader, GameType game)
+        => Read(reader, game);
+
+    public override void Read(BinaryObjectReader reader, GameType game)
+    {
+        Field00 = reader.Read<int>();
+        Field04 = reader.Read<int>();
+        Field08 = reader.Read<float>();
+        Field0C = reader.Read<float>();
+        Field10 = reader.Read<int>();
+        Field14 = reader.Read<int>();
+        Field18 = reader.Read<int>();
+        Field1C = reader.Read<int>();
+    }
+
+    public override void Write(BinaryObjectWriter writer, GameType game)
+    {
+        writer.Write(Field00);
+        writer.Write(Field04);
+        writer.Write(Field08);
+        writer.Write(Field0C);
+        writer.Write(Field10);
+        writer.Write(Field14);
+        writer.Write(Field18);
+        writer.Write(Field1C);
+    }
+
+    public override int GetTypeID(GameType game) { return (int)ParameterType.CameraShake; }
+}
