+++ Source/SharpNeedle/SonicTeam/DiEvent/Parameters/CullDisabledParam.cs
@@ -1,0 +1,14 @@
+namespace SharpNeedle.SonicTeam.DiEvent;
+
+public class CullDisabledParam : BaseParam
+{
+    public CullDisabledParam() { }
+    public CullDisabledParam(BinaryObjectReader reader, GameType game)
+        => Read(reader, game);
+
+    public override void Read(BinaryObjectReader reader, GameType game) { }
+
+    public override void Write(BinaryObjectWriter writer, GameType game) { }
+
+    public override int GetTypeID(GameType game) { return (int)ParameterType.CullDisabled; }
+}
