+++ Source/SharpNeedle/SonicTeam/DiEvent/Parameters/EffectParam.cs
@@ -1,0 +1,55 @@
+namespace SharpNeedle.SonicTeam.DiEvent;
+
+public class EffectParam : BaseParam
+{
+    public Matrix4x4 LocalTransform { get; set; }
+    public int Field40 { get; set; }
+    public string Name { get; set; }
+    public int Field84 { get; set; }
+    public int Field88 { get; set; }
+    public int Field8C { get; set; }
+    public int Field90 { get; set; }
+    public int Field94 { get; set; }
+    public int Field98 { get; set; }
+    public int Field9C { get; set; }
+    public int FieldA0 { get; set; }
+    public float[] AnimationData { get; set; } = new float[128];
+
+    public EffectParam() { }
+    public EffectParam(BinaryObjectReader reader, GameType game)
+        => Read(reader, game);
+
+    public override void Read(BinaryObjectReader reader, GameType game)
+    {
+        LocalTransform = reader.Read<Matrix4x4>();
+        Field40 = reader.Read<int>();
+        Name = reader.ReadDiString(64);
+        Field84 = reader.Read<int>();
+        Field88 = reader.Read<int>();
+        Field8C = reader.Read<int>();
+        Field90 = reader.Read<int>();
+        Field94 = reader.Read<int>();
+        Field98 = reader.Read<int>();
+        Field9C = reader.Read<int>();
+        FieldA0 = reader.Read<int>();
+        reader.ReadArray<float>(128, AnimationData);
+    }
+
+    public override void Write(BinaryObjectWriter writer, GameType game)
+    {
+        writer.Write(LocalTransform);
+        writer.Write(Field40);
+        writer.WriteDiString(Name);
+        writer.Write(Field84);
+        writer.Write(Field88);
+        writer.Write(Field8C);
+        writer.Write(Field90);
+        writer.Write(Field94);
+        writer.Write(Field98);
+        writer.Write(Field9C);
+        writer.Write(FieldA0);
+        writer.WriteArrayFixedLength(AnimationData, 128);
+    }
+
+    public override int GetTypeID(GameType game) { return (int)ParameterType.Effect; }
+}
