+++ Source/SharpNeedle/SonicTeam/DiEvent/Parameters/GameCameraParam.cs
@@ -1,0 +1,97 @@
+namespace SharpNeedle.SonicTeam.DiEvent;
+
+public class GameCameraParam : BaseParam
+{
+    public int Field00 { get; set; }
+    public int Field04 { get; set; }
+    public int Field08 { get; set; }
+    public int Field0C { get; set; }
+    public float Field10 { get; set; }
+    public float Field14 { get; set; }
+    public float Field18 { get; set; }
+    public int Field1C { get; set; }
+    public int Field20 { get; set; }
+    public int Field24 { get; set; }
+    public float Field28 { get; set; }
+    public float Field2C { get; set; }
+    public float Field30 { get; set; }
+    public int Field34 { get; set; }
+    public int Field38 { get; set; }
+    public int Field3C { get; set; }
+    public int Field40 { get; set; }
+    public float Field44 { get; set; }
+    public int Field48 { get; set; }
+    public int Field4C { get; set; }
+    public float NearCullingPlane { get; set; }
+    public float FarCullingPlane { get; set; }
+    public float Field58 { get; set; }
+    public int Field5C { get; set; }
+    public int Field60 { get; set; }
+    public int Field64 { get; set; }
+
+    public GameCameraParam() { }
+    public GameCameraParam(BinaryObjectReader reader, GameType game)
+        => Read(reader, game);
+
+    public override void Read(BinaryObjectReader reader, GameType game)
+    {
+        Field00 = reader.Read<int>();
+        Field04 = reader.Read<int>();
+        Field08 = reader.Read<int>();
+        Field0C = reader.Read<int>();
+        Field10 = reader.Read<float>();
+        Field14 = reader.Read<float>();
+        Field18 = reader.Read<float>();
+        Field1C = reader.Read<int>();
+        Field20 = reader.Read<int>();
+        Field24 = reader.Read<int>();
+        Field28 = reader.Read<float>();
+        Field2C = reader.Read<float>();
+        Field30 = reader.Read<float>();
+        Field34 = reader.Read<int>();
+        Field38 = reader.Read<int>();
+        Field3C = reader.Read<int>();
+        Field40 = reader.Read<int>();
+        Field44 = reader.Read<float>();
+        Field48 = reader.Read<int>();
+        Field4C = reader.Read<int>();
+        NearCullingPlane = reader.Read<float>();
+        FarCullingPlane = reader.Read<float>();
+        Field58 = reader.Read<float>();
+        Field5C = reader.Read<int>();
+        Field60 = reader.Read<int>();
+        Field64 = reader.Read<int>();
+    }
+
+    public override void Write(BinaryObjectWriter writer, GameType game)
+    {
+        writer.Write(Field00);
+        writer.Write(Field04);
+        writer.Write(Field08);
+        writer.Write(Field0C);
+        writer.Write(Field10);
+        writer.Write(Field14);
+        writer.Write(Field18);
+        writer.Write(Field1C);
+        writer.Write(Field20);
+        writer.Write(Field24);
+        writer.Write(Field28);
+        writer.Write(Field2C);
+        writer.Write(Field30);
+        writer.Write(Field34);
+        writer.Write(Field38);
+        writer.Write(Field3C);
+        writer.Write(Field40);
+        writer.Write(Field44);
+        writer.Write(Field48);
+        writer.Write(Field4C);
+        writer.Write(NearCullingPlane);
+        writer.Write(FarCullingPlane);
+        writer.Write(Field58);
+        writer.Write(Field5C);
+        writer.Write(Field60);
+        writer.Write(Field64);
+    }
+
+    public override int GetTypeID(GameType game) { return (int)ParameterType.GameCamera; }
+}
